requirements.txt
streamlit>=1.28.0
ultralytics>=8.0.0
opencv-python>=4.6.0
Pillow>=9.0.0